test/t: Always rebuild cirun, but only do it once per test suite invocation

+++ test/t/lib.bash
@@ -8,24 +8,28 @@ tmp_dir=/tmp/cirun-test
 mkdir -p "$tmp_dir"
 
 # Build
-(
-	flock 9
-	cd ../..
-	if [[ ! -x ./cirun ]]
-	then
-		dub build
-	fi
-) 9>> "$tmp_dir"/build.lock
+if [[ ! -v CIRUN_BUILT ]]
+then
+	(
+		flock 9
+		cd ../..
+		dub build -q
+	) 9>> "$tmp_dir"/build.lock
+	export CIRUN_BUILT=1
+fi
 cirun=$PWD/../../cirun
 
 test_name=$(basename "$0" .sh)
-test_dir=$tmp_dir/$test_name
+if [[ "$test_name" != all ]]
+then
+	test_dir=$tmp_dir/$test_name
 
-rm -rf "$test_dir"
-mkdir "$test_dir"
-cd "$test_dir"
+	rm -rf "$test_dir"
+	mkdir "$test_dir"
+	cd "$test_dir"
 
-{
-	printf 'dataDir = cirun-data\n'
-	printf 'workDir = cirun-work\n'
-} > cirun.conf
+	{
+		printf 'dataDir = cirun-data\n'
+		printf 'workDir = cirun-work\n'
+	} > cirun.conf
+fi
